Increase the speed of players

+++ Shared/LuceRPG.Models/WorldObject.fs
@@ -73,7 +73,7 @@ module WorldObject =
     /// Time taken by the object to move one square
     let travelTime (obj: Payload): int64 =
         match obj.t with
-        | Type.Player _ -> System.TimeSpan.FromMilliseconds(float(300)).Ticks
+        | Type.Player _ -> System.TimeSpan.FromMilliseconds(float(250)).Ticks
         | _ -> 0L
 
     let isPlayer (obj: Model): bool =
